requirements.txt
ultralytics
fastapi
uvicorn
python-multipart
pillow
httpx
pytest